fix(workspace): invalidate descendants on dir moves; extract nodeWatcher module
Round-2 review findings:

- Directory moves stranded descendant caches: the synthetic rename
  carries only the top-level from/to, but file/stat/tree queries are
  keyed by full path — an editor open on a file under the moved folder
  kept stale content on a dead key (the suppressed echo used to do this
  cleanup per child). The moved handler now also invalidates everything
  cached under the old prefix via a predicate filter — only ACTIVE
  descendant queries (open editors) refetch, and file moves match
  nothing. The batch gained enqueueFilter so predicate invalidations
  ride the same 25ms coalescing as key-based ones.

- Extract nodeWatcher.ts: the watcher subsystem (echo suppression,
  size guard, chokidar wiring) had grown to ~60% of
  createNodeWorkspace.ts. Now 154-line fs implementation +
  245-line watcher module, each cohesive. No behavior change.

- Drop the last 'as const' in fsEvents.ts in favor of an explicit
  WorkspaceWatcherReadiness annotation.
