+++ packages/agent/src/server/workspace/createNodeWorkspace.ts
@@ -1,258 +1,17 @@
 import { lstat, mkdir, readdir, readFile, rename, rm, stat, unlink, writeFile } from 'node:fs/promises'
-import { dirname, join, relative, resolve, sep } from 'node:path'
-import chokidar, { type FSWatcher } from 'chokidar'
+import { dirname, resolve } from 'node:path'
 
 import type { WorkspaceRuntimeContext } from '../../shared/runtime'
-import type {
-  Workspace,
-  WorkspaceChangeEvent,
-  WorkspaceWatcher,
-  WorkspaceWatcherReadiness,
-} from '../../shared/workspace'
+import type { Workspace } from '../../shared/workspace'
 import {
   assertRealPathWithinWorkspace,
   ensureExistingWorkspacePath,
   ensureWritableWorkspacePath,
   validatePath,
 } from './paths'
-import { isIgnoredDirName } from './ignore'
+import { createNodeWatcher, toPosixRel, type NodeWorkspaceWatcher } from './nodeWatcher'
 
 const EPERM_CODE = 'EPERM'
-
-function shouldIgnoreWatchPath(root: string, path: string): boolean {
-  const relPath = relative(root, path)
-  const parts = relPath.split(sep)
-  return parts.some((part) => isIgnoredDirName(part))
-}
-
-/** Workspace-relative path with POSIX separators — the wire format for
- * every `WorkspaceChangeEvent.path`. */
-function toPosixRel(root: string, absPath: string): string {
-  return relative(root, absPath).split(sep).join('/')
-}
-
-/**
- * Watcher with one host-facing extra: `emitRename`. Chokidar has no
- * rename primitive — a moved directory surfaces as an `unlink` per old
- * path plus an `add` per new path, thousands of events for a big tree.
- * When the host performed the rename itself (the files/move API), it
- * announces it here: subscribers get ONE synthetic `rename` event and
- * the watcher absorbs exactly the echo that rename predicts —
- * `unlink`/`unlinkDir` under the old prefix, `add`/`addDir` under the
- * new one — while it lasts (sliding idle window, hard-capped). Other
- * event kinds under either prefix (a real edit or delete inside the
- * moved folder) flow through untouched.
- */
-interface NodeWorkspaceWatcher extends WorkspaceWatcher {
-  emitRename(fromRel: string, toRel: string): void
-}
-
-/** Echo absorption stops this long after the last suppressed event. */
-const RENAME_ECHO_IDLE_MS = 1_000
-/** Absolute ceiling so a prefix can never be muted indefinitely. */
-const RENAME_ECHO_MAX_MS = 30_000
-
-/** Raw chokidar event names — suppression matches on these, not on the
- * collapsed `WorkspaceChangeEvent.op` (add and change both map to
- * `write`, but only add is rename echo). */
-type ChokidarEventKind = 'add' | 'addDir' | 'change' | 'unlink' | 'unlinkDir'
-
-interface EchoSuppression {
-  prefix: string
-  /** Event kinds this prefix absorbs — the rename's predicted echo. */
-  kinds: ReadonlySet<ChokidarEventKind>
-  idleDeadline: number
-  hardDeadline: number
-}
-
-const RENAME_ECHO_FROM_KINDS: ReadonlySet<ChokidarEventKind> = new Set(['unlink', 'unlinkDir'])
-const RENAME_ECHO_TO_KINDS: ReadonlySet<ChokidarEventKind> = new Set(['add', 'addDir'])
-
-/**
- * Watching a workspace means chokidar enumerates (and on some
- * platforms opens an fd for) every non-ignored entry under the root.
- * Pointed at a huge directory (~ or a data dump) that is a multi-
- * minute scan, unbounded heap growth, and on macOS without fsevents an
- * silent EMFILE storm. Refuse to watch past this many entries — the
- * workspace stays fully usable, clients just fall back to non-live
- * behavior and the user gets told to start in a subfolder.
- */
-const DEFAULT_MAX_WATCHED_ENTRIES = 50_000
-
-function maxWatchedEntries(): number {
-  const raw = process.env.BORING_MAX_WATCHED_ENTRIES
-  if (!raw) return DEFAULT_MAX_WATCHED_ENTRIES
-  const n = Number.parseInt(raw, 10)
-  return Number.isFinite(n) && n > 0 ? n : DEFAULT_MAX_WATCHED_ENTRIES
-}
-
-/**
- * Bounded entry count: readdir-only walk (no stat, no fds held) that
- * skips the same dirs the watcher ignores and exits as soon as the cap
- * is crossed — on an over-sized workspace this returns in the time it
- * takes to enumerate cap+1 entries, not the whole tree.
- */
-async function countWatchableEntries(root: string, cap: number): Promise<number> {
-  let count = 0
-  const stack: string[] = [root]
-  while (stack.length > 0) {
-    const dir = stack.pop()!
-    let entries
-    try {
-      entries = await readdir(dir, { withFileTypes: true })
-    } catch {
-      continue // unreadable dir — chokidar would skip it too
-    }
-    for (const entry of entries) {
-      count += 1
-      if (count > cap) return count
-      // Mirrors the watch config: don't descend ignored names, don't
-      // follow symlinks.
-      if (entry.isDirectory() && !isIgnoredDirName(entry.name)) {
-        stack.push(join(dir, entry.name))
-      }
-    }
-  }
-  return count
-}
-
-/**
- * One chokidar instance per workspace root, fanned out to N
- * subscribers. Created lazily on first `watch()` call so unit tests
- * and watch-free hosts pay nothing.
- */
-function createNodeWatcher(root: string): NodeWorkspaceWatcher {
-  const listeners = new Set<(e: WorkspaceChangeEvent) => void>()
-  const suppressions: EchoSuppression[] = []
-  let fsw: FSWatcher | null = null
-  let closed = false
-  let readiness: Promise<WorkspaceWatcherReadiness> | null = null
-
-  const startFsw = (): void => {
-    if (fsw || closed) return
-    fsw = chokidar.watch(root, {
-      ignored: (path) => shouldIgnoreWatchPath(root, path),
-      ignoreInitial: true,
-      persistent: true,
-      followSymlinks: false,
-      // Avoid chokidar's awaitWriteFinish polling loop in long-lived app
-      // servers. Consumers already reconcile by mtime after invalidation.
-      // No native renames from chokidar — `unlinkDir`/`addDir`/
-      // `unlink`/`add` are the primitives we get. We surface them as
-      // separate `unlink` + `write`/`mkdir` events. Renames the host
-      // performs itself arrive via `emitRename` instead, which also
-      // mutes this echo.
-    })
-    fsw.on('add', (p, s) => emit('add', { op: 'write', path: rel(p), mtimeMs: s?.mtimeMs }))
-    fsw.on('change', (p, s) => emit('change', { op: 'write', path: rel(p), mtimeMs: s?.mtimeMs }))
-    fsw.on('addDir', (p) => emit('addDir', { op: 'mkdir', path: rel(p) }))
-    fsw.on('unlink', (p) => emit('unlink', { op: 'unlink', path: rel(p) }))
-    fsw.on('unlinkDir', (p) => emit('unlinkDir', { op: 'unlink', path: rel(p) }))
-    let loggedError = false
-    fsw.on('error', (err) => {
-      // Watch errors are best-effort, but a silent EMFILE storm (macOS
-      // fd limit without fsevents) looked like the app hanging. Log the
-      // first one so there's a trail.
-      if (loggedError) return
-      loggedError = true
-      console.error(
-        `[boring-agent] file watcher error in ${root} (live file events may be incomplete):`,
-        err instanceof Error ? err.message : err,
-      )
-    })
-  }
-
-  // Size guard runs once, before the first chokidar instance: counting
-  // entries is a cheap readdir walk with an early exit, while letting
-  // chokidar loose on an over-sized tree can OOM the process.
-  const ensureFsw = (): Promise<WorkspaceWatcherReadiness> => {
-    if (readiness) return readiness
-    readiness = (async (): Promise<WorkspaceWatcherReadiness> => {
-      const cap = maxWatchedEntries()
-      const count = await countWatchableEntries(root, cap)
-      if (closed) return { ok: false, reason: 'closed' }
-      if (count > cap) {
-        const message =
-          `workspace at ${root} has more than ${cap} entries — file watching disabled. `
-          + `Start boring-ui in a smaller subfolder for live file updates, `
-          + `or raise BORING_MAX_WATCHED_ENTRIES.`
-        console.error(`[boring-agent] ${message}`)
-        return { ok: false, reason: 'workspace_too_large', message }
-      }
-      startFsw()
-      return { ok: true }
-    })()
-    return readiness
-  }
-
-  const rel = (abs: string): string => toPosixRel(root, abs)
-
-  const fanout = (event: WorkspaceChangeEvent) => {
-    if (closed) return
-    if (event.path === '' || event.path.startsWith('..')) return
-    for (const l of [...listeners]) {
-      try { l(event) } catch { /* one bad listener doesn't kill the chain */ }
-    }
-  }
-
-  const isSuppressedEcho = (kind: ChokidarEventKind, path: string): boolean => {
-    const now = Date.now()
-    for (let i = suppressions.length - 1; i >= 0; i--) {
-      const s = suppressions[i]!
-      if (now > s.idleDeadline || now > s.hardDeadline) {
-        suppressions.splice(i, 1)
-        continue
-      }
-      if (s.kinds.has(kind) && (path === s.prefix || path.startsWith(`${s.prefix}/`))) {
-        s.idleDeadline = Math.min(now + RENAME_ECHO_IDLE_MS, s.hardDeadline)
-        return true
-      }
-    }
-    return false
-  }
-
-  // Chokidar events go through the suppression filter; synthetic
-  // events from `emitRename` bypass it via `fanout` directly.
-  const emit = (kind: ChokidarEventKind, event: WorkspaceChangeEvent) => {
-    if (isSuppressedEcho(kind, event.path)) return
-    fanout(event)
-  }
-
-  return {
-    subscribe(listener) {
-      if (closed) return () => {}
-      listeners.add(listener)
-      void ensureFsw()
-      return () => {
-        listeners.delete(listener)
-      }
-    },
-    whenReady() {
-      return ensureFsw()
-    },
-    emitRename(fromRel, toRel) {
-      if (closed) return
-      const now = Date.now()
-      const window = {
-        idleDeadline: now + RENAME_ECHO_IDLE_MS,
-        hardDeadline: now + RENAME_ECHO_MAX_MS,
-      }
-      suppressions.push(
-        { prefix: fromRel, kinds: RENAME_ECHO_FROM_KINDS, ...window },
-        { prefix: toRel, kinds: RENAME_ECHO_TO_KINDS, ...window },
-      )
-      fanout({ op: 'rename', path: toRel, oldPath: fromRel })
-    },
-    close() {
-      if (closed) return
-      closed = true
-      listeners.clear()
-      suppressions.length = 0
-      fsw?.close().catch(() => {})
-      fsw = null
-    },
-  }
-}
 
 export interface CreateNodeWorkspaceOptions {
   runtimeContext?: WorkspaceRuntimeContext
